+++ tests/test_visitors/test_ast/test_statements/test_useless_node.py
@@ -1,5 +1,6 @@
 import pytest
 
+from wemake_python_styleguide.compat.constants import PY311
 from wemake_python_styleguide.violations.consistency import UselessNodeViolation
 from wemake_python_styleguide.visitors.ast.statements import (
     StatementsWithBodiesVisitor,
@@ -33,12 +34,28 @@ def wrapper():
             some_call()
 """
 
+try_star_template = """
+def wrapper():
+    for wrapping_loop in []:
+        try:
+            {0}
+        except* Exception:
+            some_call()
+"""
+
 
 @pytest.mark.parametrize('code', [
     for_template,
     while_template,
     with_template,
     try_template,
+    pytest.param(
+        try_star_template,
+        marks=pytest.mark.skipif(
+            not PY311,
+            reason='ExceptionGroup was added in python 3.11',
+        ),
+    ),
 ])
 @pytest.mark.parametrize('statement', [
     'break',
@@ -89,6 +106,13 @@ def test_useless_loop_nodes(
 
 @pytest.mark.parametrize('code', [
     try_template,
+    pytest.param(
+        try_star_template,
+        marks=pytest.mark.skipif(
+            not PY311,
+            reason='ExceptionGroup was added in python 3.11',
+        ),
+    ),
 ])
 @pytest.mark.parametrize('statement', [
     'raise TypeError()',
@@ -109,7 +133,7 @@ def test_useless_try_nodes(
     visitor.run()
 
     assert_errors(visitor, [UselessNodeViolation])
-    assert_error_text(visitor, 'try')
+    assert_error_text(visitor, 'trystar' if 'except*' in code else 'try')
 
 
 @pytest.mark.parametrize('code', [
@@ -117,6 +141,13 @@ def test_useless_try_nodes(
     while_template,
     with_template,
     try_template,
+    pytest.param(
+        try_star_template,
+        marks=pytest.mark.skipif(
+            not PY311,
+            reason='ExceptionGroup was added in python 3.11',
+        ),
+    ),
 ])
 @pytest.mark.parametrize('statement', [
     'some_var = 1',