+++ wemake_python_styleguide/compat/functions.py
@@ -1,7 +1,7 @@
 import ast
 from typing import List, Union
 
-from wemake_python_styleguide.compat.types import AnyAssignWithWalrus
+from wemake_python_styleguide.types import AnyAssignWithWalrus
 
 
 def get_assign_targets(